Remove HttpServer and HttpServerAdapter in favor of a simple dispatch method
Today we have quite some abstractions that are essentially providing a simple
dispatch method to the plugins defining a `HttpServerTransport`. This commit
removes `HttpServer` and `HttpServerAdaptor` and introduces a simple `Dispatcher` functional
interface that deledgaed to `RestController` by default.

Relates to #18482
