+++ core/src/main/java/org/elasticsearch/common/settings/SettingsModule.java
@@ -54,6 +54,7 @@ public class SettingsModule implements Module {
     private final Logger logger;
     private final IndexScopedSettings indexScopedSettings;
     private final ClusterSettings clusterSettings;
+    private final SettingsFilter settingsFilter;
 
     public SettingsModule(Settings settings, Setting<?>... additionalSettings) {
         this(settings, Arrays.asList(additionalSettings), Collections.emptyList());
@@ -137,12 +138,13 @@ public class SettingsModule implements Module {
         final Predicate<String> acceptOnlyClusterSettings = TRIBE_CLIENT_NODE_SETTINGS_PREDICATE.negate();
         clusterSettings.validate(settings.filter(acceptOnlyClusterSettings));
         validateTribeSettings(settings, clusterSettings);
+        this.settingsFilter = new SettingsFilter(settings, settingsFilterPattern);
      }
 
     @Override
     public void configure(Binder binder) {
         binder.bind(Settings.class).toInstance(settings);
-        binder.bind(SettingsFilter.class).toInstance(new SettingsFilter(settings, settingsFilterPattern));
+        binder.bind(SettingsFilter.class).toInstance(settingsFilter);
         binder.bind(ClusterSettings.class).toInstance(clusterSettings);
         binder.bind(IndexScopedSettings.class).toInstance(indexScopedSettings);
     }
@@ -218,4 +220,6 @@ public class SettingsModule implements Module {
     public ClusterSettings getClusterSettings() {
         return clusterSettings;
     }
+
+    public SettingsFilter getSettingsFilter() { return settingsFilter; }
 }
