+++ core/src/main/java/org/elasticsearch/http/HttpServerTransport.java
@@ -21,6 +21,9 @@ package org.elasticsearch.http;
 
 import org.elasticsearch.common.component.LifecycleComponent;
 import org.elasticsearch.common.transport.BoundTransportAddress;
+import org.elasticsearch.common.util.concurrent.ThreadContext;
+import org.elasticsearch.rest.RestChannel;
+import org.elasticsearch.rest.RestRequest;
 
 public interface HttpServerTransport extends LifecycleComponent {
 
@@ -33,6 +36,15 @@ public interface HttpServerTransport extends LifecycleComponent {
 
     HttpStats stats();
 
-    void httpServerAdapter(HttpServerAdapter httpServerAdapter);
-
+    @FunctionalInterface
+    interface Dispatcher {
+        /**
+         * Dispatches the {@link RestRequest} to the relevant request handler or responds to the given rest channel directly if
+         * the request can't be handled by any request handler.
+         * @param request the request to dispatch
+         * @param channel the response channel of this request
+         * @param threadContext the nodes thread context
+         */
+        void dispatch(RestRequest request, RestChannel channel, ThreadContext threadContext);
+    }
 }
