+++ core/src/main/java/org/elasticsearch/plugins/NetworkPlugin.java
@@ -69,8 +69,11 @@ public interface NetworkPlugin {
      * See {@link org.elasticsearch.common.network.NetworkModule#HTTP_TYPE_SETTING} to configure a specific implementation.
      */
     default Map<String, Supplier<HttpServerTransport>> getHttpTransports(Settings settings, ThreadPool threadPool, BigArrays bigArrays,
-            CircuitBreakerService circuitBreakerService, NamedWriteableRegistry namedWriteableRegistry,
-            NamedXContentRegistry xContentRegistry, NetworkService networkService) {
+                                                                         CircuitBreakerService circuitBreakerService,
+                                                                         NamedWriteableRegistry namedWriteableRegistry,
+                                                                         NamedXContentRegistry xContentRegistry,
+                                                                         NetworkService networkService,
+                                                                         HttpServerTransport.Dispatcher dispatcher) {
         return Collections.emptyMap();
     }
 }
